third_party/WebKit/LayoutTests/http/tests/websocket/url-no-trailing-slash.html

<!DOCTYPE html>
<script src="/js-test-resources/js-test.js"></script>
<script src="resources/get-request-header.js"></script>
<script>
description("URL that doesn't have trailing slash should not emit empty Request-URI.");

window.jsTestIsAsync = true;

var url = "ws://127.0.0.1:8880";
var host;

var ws = new WebSocket(url);
ws.onmessage = function (evt) {
    host = getRequestHeaderFromEvent(evt, "host");
    shouldBeEqualToString("host", "127.0.0.1:8880");
    finishJSTest();
};
ws.onclose = function (evt) {
    testFailed("Unexpected close event " + evt);
    finishJSTest();
};

</script>
