third_party/blink/web_tests/http/tests/websocket/url-no-trailing-slash-expected.txt

URL that doesn't have trailing slash should not emit empty Request-URI.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS host is "127.0.0.1:8880"
PASS successfullyParsed is true

TEST COMPLETE
